Fix TypeError when LSP does logger->info
There is like basically one place where the LSP server does
`logger->info(...)`, and it's for printing a message when ignoring a
request because the file is `# typed: false`.

Before this commit, the behavior was to treat any print to stdout or
stderr as an error, and throw away the `sorbet` singleton, which which
would then cause a `TypeError` somewhere down the line, because
`sorbet.Pointer_stringify` was `(null).Pointer_stringify`.

Instead of aborting on print, now we just print on print.