Use `.then` instead of `onRuntimeInitialized`
From my reading of the generated `.js` code, it looks like the effect of
setting `onRuntimeInitialized` is the same as calling `.then` in the old
version of the code, but in the new version of the code you have to use
`.then` because the result of the `Sorbet()` constructor is a promise,
not a module. Maybe?